requirements.txt
geopandas
lightning
matplotlib
numpy
opencv-python
pillow
pycocotools
python-box
rasterio
segmentation-models-pytorch
shapely
torch
torchvision
tqdm
tensorboard